Start work on Instance/Discrete.v
This is a module which, given a set of objects, and a list of associations
(specified as a list of pairs), can be used to construct a diagram into
another category with more interesting objects and arrows.

Whether I continue to use sets over decidable types, or just drop to
lists (ignoring duplicates, since they won't increase the number of objects in
the discrete category), remains to be decided.