Play with the idea of reifying "Valid" terms, but this is far too slow
We end up adding back in tons of typing information, in order to simplify the
proofs, but really the possibility of failure is OK, is just means there are
solvable things that the tactic might not solve. I'm committing this just for
archival purposes, but will rip out the new code shortly.