Avoid `lia` in Lib/Equality.v for coq v8.10
In coq v8.10 `lia` isn’t yet strong enough to solve the proof in the
given file, so we have to do the proof manually.
We can’t use `omega` here, which would be able to solve the proof in
v8.10, because `omega` isn’t available anymore in newer versions of coq.