Defined a new Tactics file for some automation. Defined a new "local" product of objects. Defined a notion of a cone from c to F. Defined a notion of "Universal property."
Began definition of UniversalProperty file.

Fixed universe bug in Structure/Cone.v by adding universe annotations everywhere. Marked Two_Discrete as a Canonical structure.

Made progress on the universal property of the Cartesian product.

Continuing development of the proof that being a categorical product is a universal property.

Added some new lemmas to Structure/Cartesian.v and Theory/Isomorphism.v to assist in these changes.

Marked instance as export

Made progress on the proof that the categorical product is a universal property.

Finished the proof that Cartesian products are a universal property.
moved it into its own file.

Defined an equivalence of limit cones to be an equivalence of the underlying cones.
Added a new constructor for natural transformations which allows the user to omit the redundant symmetry data.
Continuing development of the proof that a limit is a universal property.

Added a new lemma to UniversalProperty.v which allows one to rephrase
what it means for a functor to be representable in terms of the Yoneda lemma.

Instead of saying "there is an isomorphism Hom(-,c) \cong F",
it is equivalent to say "there exists an element of F(c) such that
the induced natural transformation Hom(-,c) => F is a natural isomorphism."

We prove that there is a bijection between isomorphisms Hom(-,c) \cong F
and elements of F(c) such that the induced natural transformation is an isomorphism.

In Isomorphism.v, an "IsIsomorphism (f)" definition added.

Finished the proof that the limit is a universal property.

Fixed typo.

Eliminated a very convoluted programming pattern.
(I thought it was necessary but it turned out not to be)

Same as previous commit, cleaning up the code by
scrapping an unnecessarily bureaucratic style

Minor bugfix.

Proved that an object satisfying a universal property is unique up to
isomorphism.

It would be desirable to prove that this isomorphism "respects"
the universal property.

Outlined a proof of the theorem that the canonical isomorphism between objects
satisfying a universal property is the unique isomorphism respecting the structure.

Finished the main theorem about universal properties: if P is a universal property
and c is an object with property P, c is unique up to *unique* isomorphism respecting the
structure of P. E.g., if z, z' are both Cartesian products of x, y, then there is a unique
	 morphism z -> z' commuting with the projection maps.

Added a few helper lemmas to Theory/Isomorphisms.
The lemmas about properness are slightly redundant because they are automatically generated
by Program Definition in the Construction/Groupoid file, but the rewriting tactic does
not seem to automatically unify "@compose (Groupoid C)" with "@iso_compose C" and so the redundant
lemmas are convenient.

Fixed typo.

Proved that being a universal arrow is a universal property.

Added Structure/UniversalProperty/Universal to _CoqProject.