Adapt to Coq PR #18845 fixing an issue with maximal implicit arguments for notations.
The new (originally intended) behaviour for selecting maximal implicit
arguments of a notation for an applied constant in the presence of
multiple signatures of implicit arguments. Such notations now behave
like ordinary constants or abbreviations for non applied constants,
that is, they complete the application with as much implicit arguments
as possible.

This impacts the notation "naturality[ x ] := (@naturality _ _ _ _ x)"
which is now resolved with 3 trailing implicit arguments, making
tactics such as a "srewrite (naturality[foo])" failing with unresolved
arguments.

It happens that the "naturality" constant mostly uses one signature of
implicit arguments. So, the adaptation is to restrict "naturality" to
only this signature.